[REFACTOR] #119 팔로우, 스크랩, 공감 api 예외처리 추가

Closes #119

See merge request s08-webmobile2-sub2/S08P12A709!131

Former-commit-id: 094a30f24abacd4cb146e5307c49b02c70aab631

+++ backend/mindderStart/src/main/java/com/ssafy/mindder/scraps/controller/ScrapsController.java
@@ -46,7 +46,9 @@ public class ScrapsController {
 		logger.debug("scrapAdd - 호출 : " + feedIdx);
 		try {
 			int userIdx = jwtService.getUserIdx(accessToken);
-
+			if (scrapsService.findScrap(userIdx, feedIdx) != null) {
+				return ApiResponse.error(ErrorCode.VALIDATION_SCRAP_EXCEPTION);
+			}
 			scrapsService.addScrap(userIdx, feedIdx);
 			return ApiResponse.success(SuccessCode.CREATE_SCRAP);
 		} catch (Exception e) {
@@ -64,7 +66,9 @@ public class ScrapsController {
 		logger.debug("scrapRemove - 호출 : " + feedIdx);
 		try {
 			int userIdx = jwtService.getUserIdx(accessToken);
-
+			if (scrapsService.findScrap(userIdx, feedIdx) == null) {
+				return ApiResponse.error(ErrorCode.NOT_FOUND_SCRAP_EXCEPTION);
+			}
 			scrapsService.removeScrap(userIdx, feedIdx);
 			return ApiResponse.success(SuccessCode.DELETE_SCRAP);
 		} catch (Exception e) {
